upd schedule sql compiled upd schedule files compiled

+++ src/main/java/tw/idv/cha102/g7/schedule/entity/ScheduleReport.java
@@ -33,5 +33,5 @@ public class ScheduleReport {
     private Integer hostId;
 
     @Column(name = "schrep_sta")
-    private Integer schRepSta;
+    private Short schRepSta;
 }
